feat: extract the thread op as a separate component
抽取帖子操作栏为单独的组件
